Time-Stamp Microservice Test 4

+++ index.js
@@ -22,7 +22,7 @@ app.get("/", function (req, res) {
 // your first API endpoint... 
 app.get("/api/:date", function (req, res) {
   const currentTime = new Date()
-  res.json({unix: parseInt(currentTime)});
+  res.json({unix: parseFloat(currentTime)});
 });
